requirements.txt
#install with "pip3 install -r requirements.txt"

PyGithub
Flask
github-flask